requirements.txt
asgiref==3.9.1
Django==4.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
gunicorn==20.1.0
whitenoise==6.5.0
# psycopg[binary]>=3.1.0
